cnage task3

+++ src/test/java/edu/hw4/AnimalFunctionsTest.java
@@ -1,6 +1,7 @@
 package edu.hw4;
 
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -8,6 +9,7 @@ import java.util.Set;
 import java.util.TreeSet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -65,10 +67,11 @@ public class AnimalFunctionsTest {
     @Test
     @DisplayName("Testing task3")
     void testCountAnimalsByTypes() {
-        Map<Animal.Type, Integer> map = new HashMap<>();
-        map.put(Animal.Type.DOG, 2);
-        map.put(Animal.Type.CAT, 2);
-        map.put(Animal.Type.FISH, 1);
+        Map<Animal.Type, Long> map = new HashMap<>();
+        map.put(Animal.Type.DOG, 2L);
+        map.put(Animal.Type.CAT, 2L);
+        map.put(Animal.Type.FISH, 1L);
+
         assertEquals(map, AnimalFunctions.countAnimalsByTypes(
             List.of(
                 new Animal("Dog", Animal.Type.DOG, Animal.Sex.F, 2, 43, 5, false),
